% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Khan_Tier.R
\name{add_item_to_tier_list}
\alias{add_item_to_tier_list}
\title{Add an Item to the Tier List}
\usage{
add_item_to_tier_list(tier_list, item)
}
\arguments{
\item{tier_list}{A tier list object created by \code{create_tier_list}.}

\item{item}{The item to add.}
}
\value{
Updated tier list object.
}
\description{
Add an Item to the Tier List
}
\examples{
tier_list <- add_item_to_tier_list(tier_list, "Item3")
}
