% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Khan_Tier.R
\name{assign_to_tier}
\alias{assign_to_tier}
\title{Assign an Item to a Tier}
\usage{
assign_to_tier(tier_list, item, tier)
}
\arguments{
\item{tier_list}{A tier list object created by \code{create_tier_list}.}

\item{item}{The item to assign.}

\item{tier}{The tier to assign the item to.}
}
\value{
Updated tier list object.
}
\description{
Assign an Item to a Tier
}
\examples{
tier_list <- create_tier_list(items = c("Item1", "Item2"), tiers = c("S", "A"))
tier_list <- assign_to_tier(tier_list, "Item1", "S")
}
