% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ising.R
\name{compute_xi_summands_ising}
\alias{compute_xi_summands_ising}
\title{Compute xi summands for Ising model}
\usage{
compute_xi_summands_ising(X, omega, theta_threshold, emap)
}
\arguments{
\item{X}{n-by-d Ising data.}
\item{omega}{True weight matrix.}
\item{theta_threshold}{Theta parameter.}
\item{emap}{Edge-index map.}
}
\value{
Matrix \code{n} by \code{d*(d-1)/2}.
}
\description{
\code{xi_i(e) = X_{iu} X_{iv} - E[X_u X_v]} with
\code{E[X_u X_v] = tanh(theta) + w*}.
}
