% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_features.R
\name{edges_from_clique}
\alias{edges_from_clique}
\title{Edge indices of a clique}
\usage{
edges_from_clique(nodes, emap)
}
\arguments{
\item{nodes}{Integer vector of node IDs.}
\item{emap}{Edge-index map.}
}
\value{
Integer vector of edge indices.
}
\description{
Edge indices of a clique.
}
