% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_features.R
\name{edges_from_cycle}
\alias{edges_from_cycle}
\title{Edge indices of a cycle}
\usage{
edges_from_cycle(nodes, emap)
}
\arguments{
\item{nodes}{Integer vector of node IDs, in cyclic order.}
\item{emap}{Edge-index map.}
}
\value{
Integer vector of edge indices.
}
\description{
Edge indices of a cycle.
}
