% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_features.R
\name{edges_from_path}
\alias{edges_from_path}
\title{Edge indices of an open path}
\usage{
edges_from_path(nodes, emap)
}
\arguments{
\item{nodes}{Integer vector of node IDs in path order.}
\item{emap}{Edge-index map.}
}
\value{
Integer vector of edge indices.
}
\description{
Edge indices of an open path.
}
