% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_features.R
\name{edges_from_tree}
\alias{edges_from_tree}
\title{Edge indices of a 5-node tree}
\usage{
edges_from_tree(nodes, adj, emap)
}
\arguments{
\item{nodes}{Integer vector of 5 node IDs.}
\item{adj}{Full adjacency matrix.}
\item{emap}{Edge-index map.}
}
\value{
Integer vector of edge indices.
}
\description{
Edge indices of a 5-node tree.
}
