% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ising.R
\name{estimate_ising_weights}
\alias{estimate_ising_weights}
\title{Estimate Ising edge weights and edge-variance estimates}
\usage{
estimate_ising_weights(X, theta_threshold = 0.45, emap)
}
\arguments{
\item{X}{n-by-d matrix with entries in \code{{-1, +1}}.}
\item{theta_threshold}{Value of theta from equation (2.10).}
\item{emap}{Edge-index map.}
}
\value{
List with \code{w_hat} (numeric vector) and \code{var_hat} (numeric
vector), both of length \code{d*(d-1)/2}.
}
\description{
\code{w_hat_{uv}   = mean(X[, u] * X[, v]) - tanh(theta_threshold)}.
\code{var_hat_{uv} = 1 - mean(X[, u] * X[, v])^2}.
}
