% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm.R
\name{estimate_variance_ggm}
\alias{estimate_variance_ggm}
\title{Variance estimate for the debiased estimator (equation 2.7)}
\usage{
estimate_variance_ggm(theta_d)
}
\arguments{
\item{theta_d}{Debiased precision matrix.}
}
\value{
Symmetric d-by-d variance matrix (diagonal not used).
}
\description{
\code{var_mat[u, v] = theta_d[u, u] * theta_d[v, v] + theta_d[u, v]^2}.
}
