% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_index.R
\name{n_edges_total}
\alias{n_edges_total}
\title{Total number of undirected edges in a complete graph}
\usage{
n_edges_total(d)
}
\arguments{
\item{d}{Number of nodes.}
}
\value{
Integer.
}
\description{
Total number of undirected edges in a complete graph.
}
