% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Khan_Tier.R
\name{plot_tier_list_advanced}
\alias{plot_tier_list_advanced}
\title{Plot tier list with advanced features}
\arguments{
\item{tier_list}{The tier list object to be plotted}
}
\description{
This function creates an advanced plot for the tier list.
}
