% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homology.R
\name{rank_increase}
\alias{rank_increase}
\title{Increase in homology rank from adding a single edge}
\usage{
rank_increase(current_edges, new_edge, d, emap, k = 1L)
}
\arguments{
\item{current_edges}{Integer vector of current edge indices.}
\item{new_edge}{Single edge index to add.}
\item{d}{Number of nodes.}
\item{emap}{Edge-index map.}
\item{k}{Homology dimension.}
}
\value{
Non-negative integer.
}
\description{
Increase in homology rank from adding a single edge.
}
